use async queue from ext lib
