+++ ct-app/core/components/messages/message_format.py
@@ -1,12 +1,6 @@
 import re
-from asyncio import Queue
 from datetime import datetime
 
-from prometheus_client import Gauge
-
-from .singleton import Singleton
-
-QUEUE_SIZE = Gauge("ct_queue_size", "Size of the message queue")
 
 class MessageFormat:
     pattern = "{relayer} {index} {timestamp}"
@@ -42,23 +36,3 @@ class MessageFormat:
     
     def bytes(self):
         return self.format().encode()
-
-class MessageQueue(metaclass=Singleton):
-    def __init__(self):
-        self._buffer = Queue()
-
-    async def get(self) -> MessageFormat:
-        return await self.buffer.get()
-
-    @property
-    def buffer(self):
-        QUEUE_SIZE.set(self._buffer.qsize())
-        return self._buffer
-
-    @classmethod
-    def clear(cls):
-        instance = cls()
-
-        while not instance._buffer.empty():
-            instance._buffer.get_nowait()
-            instance._buffer.task_done()
