+++ ct-app/requirements.txt
@@ -1,6 +1,9 @@
 # HTTP requests
 aiohttp>=3.10.11
 
+# Async Queue
+janus==2.0.0
+
 # Support of .yaml files
 pyYAML>=6.0.1
